getting rid of `main_app` references in layouts
When devs reference `_url` and `_path` URL helpers in their layouts (shared by Koudoku) they necessarily have to prepend those with `main_app` which is a trip wire for a lot of people. 
Trying to solve that in the most elegant way possible. Inspired by http://stackoverflow.com/questions/9232175/access-main-app-helpers-when-overridings-a-rails-engine-view-layout#comment44856124_19453140